+++ maple/post.c
@@ -2156,7 +2156,7 @@ post_state(
             l=2;
 */
         if (!stat(fpath, &st))
-            prints("\n \x1b[1;37m★\x1b[m 最後編輯: %s\n \x1b[1;37m★\x1b[m 檔案大小: \x1b[1;32m%ld\x1b[m bytes", Ctime(&st.st_mtime), st.st_size);
+            prints("\n \x1b[1;37m★\x1b[m 最後編輯: %s\n \x1b[1;37m★\x1b[m 檔案大小: \x1b[1;32m%lld\x1b[m bytes", Ctime(&st.st_mtime), (long long)st.st_size);
 
     }
     else if (!(cuser.userlevel))
@@ -2190,7 +2190,7 @@ post_state(
             outs("\x1b[m");
         }
         if (!stat(fpath, &st))
-            prints("\n \x1b[1;37m★\x1b[m 最後存取: %s\n \x1b[1;37m★\x1b[m 檔案大小: \x1b[1;32m%ld\x1b[m bytes", Ctime(&st.st_mtime), st.st_size);
+            prints("\n \x1b[1;37m★\x1b[m 最後存取: %s\n \x1b[1;37m★\x1b[m 檔案大小: \x1b[1;32m%lld\x1b[m bytes", Ctime(&st.st_mtime), (long long)st.st_size);
 
     }
